refactor: discord client now in discord/mod.rs
