+++ src/discord/metrics_discord.rs
@@ -1,5 +1,7 @@
 //! Metrics constants for the discord client.
 
+pub use metrics;
+
 /// The metric name for the discord_interactions_total metric.
 pub static METRIC_DISCORD_INTERACTIONS_TOTAL: &str = "discord_interactions_total";
 /// The metric name for the discord_interactions_duration metric.
